更新文章分类功能

+++ loozb-service/src/main/java/com/loozb/core/utils/TreeUtil.java
@@ -29,7 +29,7 @@ public class TreeUtil {
         for (SysResource resource : allSysResources) {
             boolean isRoot = true;
             for (SysResource comparedOne : allSysResources) {
-                if (resource.getPid() == comparedOne.getId()) {
+                if (resource.getPid().longValue() == comparedOne.getId().longValue()) {
                     isRoot = false;
                     break;
                 }
@@ -48,7 +48,7 @@ public class TreeUtil {
         List<SysResource> children = new ArrayList<SysResource>();
 
         for (SysResource comparedOne : allSysResources) {
-            if (comparedOne.getPid() == root.getId()) {
+            if (comparedOne.getPid().longValue() == root.getId().longValue()) {
                 children.add(comparedOne);
             }
         }
